Completed Feature: Traps, Enemies and Money
Player can now build traps, with a limit on each type, each trap has a cost and can not be placed if player does not have enough money. Enemies give money when killed. Added 3 types of enemies, default, fast and tank. Added a spawner with weight on each type of enemy.